Move User and Project classes to datasets.py. Change get_project to return a Project instance. Change dataset.editor and dataset.owner to return Project or User classes when possible.
